extracted styles from template
