+++ src/app/index.module.ts
@@ -37,7 +37,7 @@ import {ResourceFetcherService} from '../components/service/resource-fetcher/res
 import {CheBranding} from '../components/branding/che-branding';
 import { RegistryCheckingService } from '../components/service/registry-checking.service';
 import { DetectSupportedBrowserService } from '../components/service/browser-detect';
-import { StorageTypeService } from '../components/service/storage-type.service';
+import { StorageTypeService } from '../components/service/storage-type/storage-type.service';
 
 // init module
 const initModule = angular.module('userDashboard', ['ngAnimate', 'ngCookies', 'ngTouch', 'ngSanitize', 'ngResource', 'ngRoute',
