+++ src/app/workspaces/workspace-details/workspace-overview/workspace-details-overview.controller.ts
@@ -18,7 +18,7 @@ import { WorkspaceDetailsService } from '../workspace-details.service';
 import { CheKubernetesNamespace } from '../../../../components/api/che-kubernetes-namespace.factory';
 import { CheDashboardConfigurationService } from '../../../../components/branding/che-dashboard-configuration.service';
 import { TogglableFeature } from '../../../../components/branding/branding.constant';
-import { StorageTypeService, StorageType } from '../../../../components/service/storage-type.service';
+import { StorageTypeService, StorageType } from '../../../../components/service/storage-type/storage-type.service';
 
 const STARTING = WorkspaceStatus[WorkspaceStatus.STARTING];
 const RUNNING = WorkspaceStatus[WorkspaceStatus.RUNNING];
