+++ src/components/service/storage-type/storage-type.styl
@@ -1,0 +1,3 @@
+.experimental-storage-type
+  font-size 0.8em
+  color #F37943
